// frontend/aada_web/src/api/generated/models/externshipRead.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * AADA LMS API
 * OpenAPI spec version: 1.0
 */
import type { ExternshipReadSiteAddress } from './externshipReadSiteAddress';
import type { ExternshipReadSupervisorName } from './externshipReadSupervisorName';
import type { ExternshipReadSupervisorEmail } from './externshipReadSupervisorEmail';
import type { ExternshipReadVerificationDocUrl } from './externshipReadVerificationDocUrl';
import type { ExternshipReadVerifiedAt } from './externshipReadVerifiedAt';

export interface ExternshipRead {
  user_id: string;
  /** @maxLength 255 */
  site_name: string;
  site_address?: ExternshipReadSiteAddress;
  supervisor_name?: ExternshipReadSupervisorName;
  supervisor_email?: ExternshipReadSupervisorEmail;
  /** @minimum 0 */
  total_hours?: number;
  verification_doc_url?: ExternshipReadVerificationDocUrl;
  verified?: boolean;
  verified_at?: ExternshipReadVerifiedAt;
  id: string;
}